python_service/requirements.txt
pandas>=2.0.0
mysql-connector-python>=8.0
